added tests for: dtlz, neighbors, weighting (in progress)

+++ src/util/dtlz.cu
@@ -33,11 +33,11 @@ __device__
 void dtlz( float *params, float *objectives, int param_size, int obj_size, int step_size) {
 
     float param_sum = 0.0;
-    for (size_t i = 0; i < param_size; i++) {
+    for (int i = 0; i < param_size; i++) {
       param_sum += params[i * step_size];
     }
 
-    for (size_t i = 0; i < obj_size; i++) {
+    for (int i = 0; i < obj_size; i++) {
       objectives[i * step_size] = param_sum * i;
     }
